requirements.txt
Flask
yfinance
Flask-Cors
firebase-admin
PyJWT
requests
gunicorn
